+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for `_partial_`
 - Fixed module resolution for `classmethod`s and `staticmethod`s
+- Refactored `PythonAnalyzer` to cache file reads
 
 ## [0.1.4]